Move checkpoint data to own file
