+++ divi/src/checkpoints.h
@@ -6,19 +6,10 @@
 #define BITCOIN_CHECKPOINTS_H
 
 #include "uint256.h"
-
-#include <map>
+#include "checkpoint_data.h"
 
 class CBlockIndex;
 
-typedef std::map<int, uint256> MapCheckpoints;
-class CCheckpointData {
-public:
-    const MapCheckpoints* mapCheckpoints;
-    int64_t nTimeLastCheckpoint;
-    int64_t nTransactionsLastCheckpoint;
-    double fTransactionsPerDay;
-};
 
 /** 
  * Block-chain checkpoints are compiled-in sanity checks.
